pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "django-sabia-auth"
version = "0.1.0"
description = "Django OAuth2 authentication backend for Sabiá (Brazil's SUS identity provider)"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "kelsoncm" }]
requires-python = ">=3.10"
dependencies = [
    "Django>=5.2,<7",
    "requests>=2.31",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "responses>=0.25",
    "black>=24.0",
    "ruff>=0.4",
    "pre-commit>=3.7",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = "-v"
pythonpath = ["."]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.setuptools.packages.find]
where = ["."]
include = ["django_sabia_auth*"]
